Add a reset button to the extension popup
