+++ addon/popup.html
@@ -13,7 +13,7 @@
       cursor: pointer;
       border-radius: 3px;
     }
-    #reload {
+    #reload,#reset {
       width: 100%;
       height: 2em;
       margin: 0.5em auto;
@@ -83,6 +83,7 @@
       <label for="retries">Detection attempts</label>
     </div>
     <button id="reload">Reload rules</button>
+    <button id="reset">Reset settings</button>
   </fieldset>
   <script src="popup.bundle.js"></script>
 </body>
