+++ addon/popup.ts
@@ -1,7 +1,7 @@
 import { BackgroundMessage } from "../lib/messages";
 import { Config } from "../lib/types";
 import { storageGet, storageRemove, storageSet } from "./mv-compat";
-import { showOptOutStatus } from "./utils";
+import { initConfig, showOptOutStatus } from "./utils";
 
 async function init() {
   const autoconsentConfig: Config = await storageGet('config');
@@ -15,6 +15,7 @@ async function init() {
   const cosmeticOffRadio = document.querySelector('input#cosmetic-off') as HTMLInputElement;
   const retriesInput = document.querySelector('input#retries') as HTMLInputElement;
   const ruleReloadButton = document.querySelector('#reload') as HTMLButtonElement;
+  const resetButton = document.querySelector('#reset') as HTMLButtonElement;
 
   // enable proceed button when necessary
 
@@ -111,6 +112,12 @@ async function init() {
     storageSet({
       rules: await res.json(),
     });
+  })
+
+  resetButton.addEventListener('click', async () => {
+    await storageRemove('config');
+    await initConfig();
+    window.close();
   })
 }
