+++ addon/utils.ts
@@ -1,4 +1,6 @@
 import { enableLogs } from "../lib/config";
+import { Config } from "../lib/types";
+import { storageGet, storageSet } from "./mv-compat";
 
 export async function showOptOutStatus(
   tabId: number,
@@ -39,4 +41,38 @@ export async function showOptOutStatus(
     tabId,
     path: icon,
   });
+}
+
+export async function initConfig() {
+  console.log('init sw');
+  const storedConfig = await storageGet('config');
+  console.log('storedConfig', storedConfig);
+  const defaultConfig: Config = {
+    enabled: true,
+    autoAction: 'optOut', // if falsy, the extension will wait for an explicit user signal before opting in/out
+    disabledCmps: [],
+    enablePrehide: true,
+    enableCosmeticRules: true,
+    detectRetries: 20,
+    isMainWorld: false,
+    prehideTimeout: 2000,
+  };
+  if (!storedConfig) {
+    console.log('init config');
+    await storageSet({
+      config: defaultConfig,
+    });
+  } else { // clean up the old stored config in case there are new fields
+    const updatedConfig: Config = structuredClone(defaultConfig);
+    for (const key of Object.keys(defaultConfig)) {
+      if (typeof storedConfig[key] !== 'undefined') {
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore - TS doesn't know that we've checked for undefined
+        updatedConfig[key] = storedConfig[key];
+      }
+    }
+    await storageSet({
+      config: updatedConfig,
+    });
+  }
 }
